requirements.txt
Flask==2.3.2
google-cloud-bigquery==3.11.4
PyJWT==2.8.0
python-dotenv==1.0.0
